Fixed logic errors within the Delete methods of Category, Manufacturer, Product, ProductType and Sale

+++ Web/DaysForGirls.Web/Areas/Administration/Controllers/SaleController.cs
@@ -189,6 +189,11 @@
         [HttpGet("/Administration/Sale/Delete/{saleId}")]
         public async Task<IActionResult> Delete(string saleId)
         {
+            if(saleId == null)
+            {
+                return BadRequest();
+            }
+
             bool isDeleted = await this.saleService.DeleteSaleById(saleId);
 
             return Redirect("/Administration/Sale/All");
